[*] fix of nonexistent dir error
git-svn-id: svn+ssh://svn.dev/svnroot/deploy@33733 16631623-8a9c-491e-a762-cca08b4963ca

+++ src/cmd.sh
@@ -88,6 +88,9 @@ clearcache() {
   execute_concurrent $packagename \
   "
     for d in $cacheglob; do
+      if [ ! -d \$d ]; then
+        continue
+      fi
       tmpdir=\`mktemp -u ${TMPDIR:-/tmp}/cache.XXXXXX\`
       sudo mv \$d \$tmpdir
       sudo rm -fr \$tmpdir
